+++ src/components/common/Router.js
@@ -1,6 +1,7 @@
 import Home from 'pages/Home';
 import Wish from 'pages/Wish';
 import Room from 'pages/Room';
+import WishList from 'pages/WishList';
 import React from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import ScrollToTop from './ScrollToTop';
@@ -11,7 +12,7 @@ function Router() {
       <ScrollToTop />
       <Routes>
         <Route path="/" element={<Home />} />
-        <Route path="/wish" element={<Wish />} />
+        <Route path="/wishlist" element={<WishList />} />
         <Route path="/room/:id" element={<Room />} />
         <Route path="/*" element={<p>Page Not Found</p>} />
       </Routes>
